+++ CHANGELOG.rst
@@ -26,7 +26,7 @@ Deprecations:
 Changes:
 ^^^^^^^^
 
-*none*
+- You can now load encrypted PKCS#8 PEM key as ``pem.Key``.
 
 
 ----