+++ CHANGELOG.rst
@@ -29,8 +29,8 @@ Deprecations:
 Changes:
 ^^^^^^^^
 
-- Added support for ``pem.OpenSSHPrivateKey`` (``OPENSSH PRIVATE KEY``). OpenSSH added a new BEGIN label when it
-  switched to a `proprietary key encoding <https://coolaj86.com/articles/the-openssh-private-key-format/>`_.
+- Added support for ``pem.OpenSSHPrivateKey`` (``OPENSSH PRIVATE KEY``).
+  OpenSSH added a new ``BEGIN`` label when it switched to a `proprietary key encoding <https://coolaj86.com/articles/the-openssh-private-key-format/>`_.
   `#39 <https://github.com/hynek/pem/pull/39>`_